+++ make_templates.sh
@@ -14,6 +14,6 @@ python tools/build_bubble_templates.py \
 	-o output_bubble_templates/
 
 sleep 0.1
-open output_bubble_templates/qc_darkness_filter/*_norm_filtered_montage.png
-sleep 0.1
 open config/bubble_templates/?.png
+sleep 0.1
+open output_bubble_templates/qc_darkness_filter/*_norm_filtered_montage.png